Переписываем bot_init.py инициализацию менеджеров БД
